requirements.txt
aws-cdk.core==1.128.0
aws-cdk.aws-s3==1.128.0
aws-cdk.aws_s3_notifications==1.128.0
aws-cdk.aws-lambda==1.128.0
aws-cdk.aws-ec2==1.128.0
aws-cdk.custom-resources==1.128.0
aws-cdk.aws-lambda-event-sources==1.128.0
aws-cdk.aws-dynamodb==1.128.0
aws-cdk.aws-iam==1.128.0
boto3
pillow
pyodata
requests
